chore: Supabase 설정 및 관련 코드 정리
- appsettings.json에 Supabase Storage 실제 설정 추가
- Transaction Repository 디렉토리 추가
- DTO/DBModel CRLF 정리
- ChatHub, TicketContext 업데이트

+++ TicketPlatFormServer/DBModel/ChatMessage.cs
@@ -18,7 +18,7 @@ public partial class ChatMessage
     /// <summary>
     /// 발신자 FK
     /// </summary>
-    public long SenderId { get; set; }
+    public int SenderId { get; set; }
 
     /// <summary>
     /// 메시지 내용
