+++ TicketPlatFormServer/DTO/Chat/CreateChatRoomReqDto.cs
@@ -2,5 +2,5 @@ namespace TicketPlatFormServer.DTO.Chat;
 
 public class CreateChatRoomReqDto
 {
-    public long TicketId { get; set; }
+    public int TicketId { get; set; }
 }
