+++ TicketPlatFormServer/DTO/Chat/NewMessageSignalDto.cs
@@ -4,7 +4,7 @@ public class NewMessageSignalDto
 {
     public long MessageId { get; set; }
     public long RoomId { get; set; }
-    public long SenderId { get; set; }
+    public int SenderId { get; set; }
     public string SenderNickname { get; set; } = null!;
     public string? SenderProfileImage { get; set; }
     public string? Message { get; set; }
